pebot-replica/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-ai>=4.14.0
datatypes-date-time>=1.0.0.a2
azure-ai-language-conversations
botbuilder-core
asyncio
aiohttp
cookiecutter==1.7.0
